Create README.md
QR Code Generator using Python. I have used "QRcode" Module here. Here we can perform this task to make a QR Code of any website or a Link.
It can be a Youtube video Link. It can also be a Link to these Github repositories.
I have imported the Module and then we have performed the task..
I have saved the output in picture format. So I can share the picture anywhere and the person only needs to scan the code.